requirements.txt
streamlit 
numpy
pandas
scikit-learn
python-dotenv
google-generativeai